+++ tiledesk-chatbot-util/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Library for Tiledesk External Chatbots, node.js
 
+
+### 0.8.8
+- added: microlang command tdAction
+- added: microlang command tdActionShowReply
+
 ### 0.8.6
 - added: microlang command tdLink
 - added: microlang command tdLinkBlank